Make the tool commands inner classes
