--- src/main/java/org/overbaard/ci/multi/repo/directory/utils/BackupMavenArtifactsToolCommand.java
@@ -1,18 +1,0 @@
-package org.overbaard.ci.multi.repo.directory.utils;
-
-import org.overbaard.ci.multi.repo.ToolCommand;
-
-/**
- * @author <a href="mailto:kabir.khan@jboss.com">Kabir Khan</a>
- */
-public class BackupMavenArtifactsToolCommand implements ToolCommand {
-    @Override
-    public String getDescription() {
-        return "Backs up the maven artifacts produced by this build";
-    }
-
-    @Override
-    public void invoke(String[] args) throws Exception {
-        BackupMavenArtifacts.backup(args);
-    }
-}
