--- src/main/java/org/overbaard/ci/multi/repo/directory/utils/OverlayBackedUpMavenArtifactsToolCommand.java
@@ -1,18 +1,0 @@
-package org.overbaard.ci.multi.repo.directory.utils;
-
-import org.overbaard.ci.multi.repo.ToolCommand;
-
-/**
- * @author <a href="mailto:kabir.khan@jboss.com">Kabir Khan</a>
- */
-public class OverlayBackedUpMavenArtifactsToolCommand implements ToolCommand {
-    @Override
-    public String getDescription() {
-        return "Overlays the maven repository with backed up maven artifacts";
-    }
-
-    @Override
-    public void invoke(String[] args) throws Exception {
-        OverlayBackedUpMavenArtifacts.overlay(args);
-    }
-}
